refactor(ExpenditureChart): unify FinancialsTrackingCard and SingleExpenditureCard into ExpenditureCard
- Add ExpenditureCard.jsx with useProjectList hook and mode='comparison'|'stacked'
- Add ExpenditureCard.module.css with shared card styles and stacked variants
- Replace FinancialsTrackingCard with thin wrapper: <ExpenditureCard mode="comparison" />
- Replace SingleExpenditureCard with thin wrapper: <ExpenditureCard mode="stacked" pieType={…} />
- Remove duplicated FinancialsTrackingCard.module.css and SingleExpenditureCard.module.css

Fixes SonarQube duplication (~40 lines); API and behavior unchanged.

Made-with: Cursor
